+++ Braintree.Tests/ResourceCollectionTest.cs
@@ -53,5 +53,54 @@ namespace Braintree.Tests
 
             Assert.AreEqual(values.Length, count);
         }
+
+        [Test]
+        [Category("Unit")]
+        public void Ids_ReturnsAllIdsInCollection()
+        {
+            string body = @"<search-results>
+                              <page-size>2</page-size>
+                              <ids type='array'>
+                                <items>0</items>
+                                <items>1</items>
+                                <items>2</items>
+                                <items>3</items>
+                                <items>4</items>
+                              </ids>
+                            </search-results>";
+
+            XmlDocument doc = new XmlDocument();
+            doc.LoadXml(body);
+            NodeWrapper xml = new NodeWrapper(doc.ChildNodes[0]);
+
+            ResourceCollection<string> resourceCollection = new ResourceCollection<string>(xml, delegate(string[] ids) {
+                return new List<string>();
+            });
+
+            List<string> assertIds = new List<string>() {"0","1","2","3","4"};
+            Assert.AreEqual(resourceCollection.Ids, assertIds);
+        }
+
+        [Test]
+        [Category("Unit")]
+        public void Ids_ReturnsEmptyListWhenNoIds()
+        {
+            string body = @"<search-results>
+                              <page-size>2</page-size>
+                              <ids type='array'>
+                              </ids>
+                            </search-results>";
+
+            XmlDocument doc = new XmlDocument();
+            doc.LoadXml(body);
+            NodeWrapper xml = new NodeWrapper(doc.ChildNodes[0]);
+
+            ResourceCollection<string> resourceCollection = new ResourceCollection<string>(xml, delegate(string[] ids) {
+                return new List<string>();
+            });
+
+            List<string> assertIds = new List<string>();
+            Assert.AreEqual(resourceCollection.Ids, assertIds);
+        }
     }
 }